+++ drift/lib/src/runtime/executor/stream_queries.dart
@@ -80,8 +80,6 @@ class StreamQueryStore {
   final StreamController<Set<TableUpdate>> _tableUpdates =
       StreamController.broadcast(sync: true);
 
-  StreamQueryStore();
-
   /// Creates a new stream from the select statement.
   Stream<List<Map<String, Object?>>> registerStream(
       QueryStreamFetcher fetcher) {