+++ drift/lib/wasm.dart
@@ -56,9 +56,12 @@ class WasmDatabase extends DelegatedDatabase {
     WasmDatabaseSetup? setup,
     IndexedDbFileSystem? fileSystem,
     bool logStatements = false,
+    bool cachePreparedStatements = true,
   }) {
     return WasmDatabase._(
-        _WasmDelegate(sqlite3, path, setup, fileSystem), logStatements);
+      _WasmDelegate(sqlite3, path, setup, fileSystem, cachePreparedStatements),
+      logStatements,
+    );
   }
 
   /// Creates an in-memory database in the loaded [sqlite3] database.
@@ -66,9 +69,12 @@ class WasmDatabase extends DelegatedDatabase {
     CommonSqlite3 sqlite3, {
     WasmDatabaseSetup? setup,
     bool logStatements = false,
+    bool cachePreparedStatements = true,
   }) {
     return WasmDatabase._(
-        _WasmDelegate(sqlite3, null, setup, null), logStatements);
+      _WasmDelegate(sqlite3, null, setup, null, cachePreparedStatements),
+      logStatements,
+    );
   }
 
   /// For an in-depth
@@ -110,8 +116,15 @@ class _WasmDelegate extends Sqlite3Delegate<CommonDatabase> {
   final IndexedDbFileSystem? _fileSystem;
 
   _WasmDelegate(
-      this._sqlite3, this._path, WasmDatabaseSetup? setup, this._fileSystem)
-      : super(setup);
+    this._sqlite3,
+    this._path,
+    WasmDatabaseSetup? setup,
+    this._fileSystem,
+    bool cachePreparedStatements,
+  ) : super(
+          setup,
+          cachePreparedStatements: cachePreparedStatements,
+        );
 
   @override
   CommonDatabase openDatabase() {
@@ -163,6 +176,8 @@ class _WasmDelegate extends Sqlite3Delegate<CommonDatabase> {
 
   @override
   Future<void> close() async {
+    await super.close();
+
     if (closeUnderlyingWhenClosed) {
       database.dispose();
       await _flush();