+++ sqlparser/lib/src/engine/sql_engine.dart
@@ -270,22 +270,18 @@ class SqlEngine {
     final node = context.root;
     node.scope = context.rootScope;
 
-    try {
-      AstPreparingVisitor(context: context).start(node);
+    AstPreparingVisitor(context: context).start(node);
 
-      node
-        ..accept(ColumnResolver(context), const ColumnResolverContext())
-        ..accept(ReferenceResolver(context), const ReferenceResolvingContext());
+    node
+      ..accept(ColumnResolver(context), const ColumnResolverContext())
+      ..accept(ReferenceResolver(context), const ReferenceResolvingContext());
 
-      final session = TypeInferenceSession(context, options);
-      final resolver = TypeResolver(session);
-      resolver.run(node);
-      context.types2 = session.results!;
+    final session = TypeInferenceSession(context, options);
+    final resolver = TypeResolver(session);
+    resolver.run(node);
+    context.types2 = session.results!;
 
-      node.acceptWithoutArg(LintingVisitor(options, context));
-    } catch (_) {
-      rethrow;
-    }
+    node.acceptWithoutArg(LintingVisitor(options, context));
   }
 }