+++ drift/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.8.1
+
+- Performance improvement: Cache and re-use prepared statements - thanks to [@davidmartos96](https://github.com/davidmartos96/)
+- Fix a deadlock after rolling back a transaction in a remote isolate.
+- Remove unintended log messages when using `connectToDriftWorker`.
+
 ## 2.8.0
 
 - Don't keep databases in an unusable state if the `setup` callback throws an